% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpRankTest.R
\name{cmpRankTest}
\alias{cmpRankTest}
\title{Composite Rank-Based Test}
\usage{
cmpRankTest(
  X1,
  X2,
  omega = 0.5,
  W = "CV",
  order = 0,
  method = "mpt",
  weight.fun = "parz",
  ...
)
}
\arguments{
\item{X1}{A numeric \eqn{(n_1 \times p)} matrix of data from sample 1.}

\item{X2}{A numeric \eqn{(n_2 \times p)} matrix of data from sample 2.}

\item{omega}{The true value of the relative effect.}

\item{W}{The lag window width. If \code{NA}, the elbow method is used to
determine the value. If \code{"CV"}, the value is selected via cross-validation
using the tuning parameters, default to \code{step = 1} and \code{fold = 10}.}

\item{order}{The order of the centering correction.
Available options are 0, 1, and 2.}

\item{method}{The scale estimation method.
Available options are "mpt" and "gct".}

\item{weight.fun}{window weight function, \code{"parz"} for Parzen weight and
\code{"trapez"} for trapezoid weight.}

\item{...}{Arguments passed to the \code{find_win} function.
Default options \code{step = 1}, \code{cv.fold = 10},
and \code{norm.type = "E"}}
}
\value{
values of test statistic, p-value and window width.
}
\description{
Kong, Villasante Tezanos and Harrar (2022): Generalized nonparametric
composite tests for high-dimensional data.
}
\examples{
# generating 2-sample data
set.seed(123)
p = 300
n1 = 60
n2 = 80
X1 = matrix(rnorm(n1*p), nrow = n1, ncol = p)
X2 = matrix(rnorm(n2*p), nrow = n2, ncol = p)

cmpRankTest(X1, X2)
cmpRankTest(X1, X2, method = "gct")


}
