% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpTest.R
\name{cmpTest}
\alias{cmpTest}
\title{Composite Test}
\usage{
cmpTest(X1, X2, W = NA, order = 0, method = "mpt", weight.fun = "parz")
}
\arguments{
\item{X1}{A numeric \eqn{(n_1 \times p)} matrix of data from sample 1.}

\item{X2}{A numeric \eqn{(n_2 \times p)} matrix of data from sample 2.}

\item{W}{The lag window width. If \code{NA}, defaults to \code{ceiling(p^(3/8))}}

\item{order}{The order of the centering correction.
Available options are 0, 1, and 2.}

\item{method}{The scale estimation method.
Available options are \code{"mpt"} and \code{"gct"}.}

\item{weight.fun}{window weight function, \code{"parz"} for Parzen weight and
\code{"trapez"} for trapezoid weight.}
}
\value{
values of test statistic, p-value and window width.
}
\description{
Gregory, Carroll, Baladandayuthapani and Lahiri (2015, JASA) and
Zhang and Wang (2020, CSDA)
}
\examples{
# generating 2-sample data
set.seed(123)
p = 300
n1 = 60
n2 = 80
X1 = matrix(rnorm(n1*p), nrow = n1, ncol = p)
X2 = matrix(rnorm(n2*p), nrow = n2, ncol = p)

cmpTest(X1, X2)
cmpTest(X1, X2, method = "gct")


}
