% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovlTest.R
\name{ovlTest}
\alias{ovlTest}
\title{Overall Mean Test}
\usage{
ovlTest(X1, X2)
}
\arguments{
\item{X1}{A numeric \eqn{(n_1 \times p)} matrix of data from sample 1.}

\item{X2}{A numeric \eqn{(n_2 \times p)} matrix of data from sample 2.}
}
\value{
values of test statistic and p-value.
}
\description{
Chen and Qin (2010, AOS): A two-sample test for high-dimensional data
with applications to gene-set testing.
}
\examples{
# generating 2-sample data
set.seed(123)
p = 300
n1 = 60
n2 = 80
X1 = matrix(rnorm(n1*p), nrow = n1, ncol = p)
X2 = matrix(rnorm(n2*p), nrow = n2, ncol = p)

ovlTest(X1, X2)

}
