% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rData.R
\name{rData}
\alias{rData}
\title{Generate Random Data}
\usage{
rData(
  p = 300,
  n = 60,
  mu = rep(0, p),
  type = "arma.norm",
  arma = list(order = c(0, 0, 0)),
  H = 0.7,
  Sigma = diag(p)
)
}
\arguments{
\item{p}{An integer specifying the number of variables (dimension).}

\item{n}{An integer specifying the number of observations (sample size).}

\item{mu}{A numeric vector of the means.}

\item{type}{A character string indicating the data generation type
(e.g., \code{"arma.norm"}, \code{"arma.t"}, \code{"arma.cauchy"}, \code{"arma.gamma"},
\code{"ld.norm"}, \code{"ld.t"}, \code{"ld.cauchy"}, \code{"ld.gamma"},
\code{"cov.norm"}, \code{"cov.t"}, \code{"cov.cauchy"}).\describe{
\item{"normal"}{Normal(mean = 0, sd = 1);}
\item{"t"}{scaled t(df = 3);}
\item{"cauchy"}{Cauchy(location = 0, scale = 0.1);}
\item{"gamma"}{centered Gamma(shape = 4, scale = 0.5).}
}}

\item{arma}{A list containing ARMA model parameters, such as
IND: \code{list(order=c(0,0,0))};
WD: \code{list(ar = c(.4, -.1), ma = c(0.2, 0.3))};
SD: \code{list(ar = 0.9)}.}

\item{H}{A numeric value representing the Hurst exponent for long-range dependence.
Default for LD: 0.7.}

\item{Sigma}{A \code{p} by \code{p} covariance matrix. Default is the identity matrix.}
}
\value{
A matrix of dimensions \code{n} by \code{p} containing the generated random data.
}
\description{
Generates simulated data from an ARMA model, a multivariate
normal distribution, or a multivariate t distribution.
}
\examples{
set.seed(123)
X1 <- rData(p = 300, n = 60)
X2 <- rData(p = 300, n = 80)

}
