% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adpTest.R
\name{adpTest}
\alias{adpTest}
\title{Adaptive Test}
\usage{
adpTest(X1, X2, W = NA, pow = c(1:6, Inf), ...)
}
\arguments{
\item{X1}{A numeric \eqn{(n_1 \times p)} matrix of data from sample 1.}

\item{X2}{A numeric \eqn{(n_2 \times p)} matrix of data from sample 2.}

\item{W}{The lag window width. If \code{NA}, the elbow method is used to
determine the value. If \code{"CV"}, the value is selected via
cross-validation using the tuning parameters \code{step} and \code{fold}.}

\item{pow}{A vector of powers used to construct the adaptive test.}

\item{...}{Arguments passed to the \code{find_win} function.
Default options \code{step = 1}, \code{cv.fold = 10},
and \code{norm.type = "E"}}
}
\value{
values of test statistic, p-value and window width.
}
\description{
XU, Lin, Wei and Pan (2016, Biometrika)
}
\examples{
# generating 2-sample data
set.seed(123)
p = 300
n1 = 60
n2 = 80
X1 = matrix(rnorm(n1*p), nrow = n1, ncol = p)
X2 = matrix(rnorm(n2*p), nrow = n2, ncol = p)

adpTest(X1, X2, W = "CV")


}
